fix: functions returning wrong values, as they were using int instead of double

+++ src/includes/raytracing/matrix/matrix_operations.h
@@ -6,7 +6,7 @@
 /*   By: gacalaza <gacalaza@student.42sp.org.br     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 14:35:20 by dmanoel-          #+#    #+#             */
-/*   Updated: 2024/05/04 22:02:16 by gacalaza         ###   ########.fr       */
+/*   Updated: 2024/05/04 22:19:34 by gacalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ void	matrix_mult_scalar(t_matrix *a, double scalar);
 void	matrix_div_scalar(t_matrix *a, double scalar);
 int		matrix_mult(t_matrix *a, t_matrix *b, t_matrix *c);
 int		matrix_transpose(t_matrix *a, t_matrix *b);
-int		matrix_determinant_4x4(t_matrix *matrix_4x4, t_matrix *result);
+double	matrix_determinant_4x4(t_matrix *matrix_4x4, t_matrix *result);
 double	cofactor_3x3(t_matrix *matrix_4x4, int row, int col);
 int		matrix_inverse_4x4(t_matrix *matrix_4x4, t_matrix *result);
